requirements.txt

aiogram==2.25.2
openai==1.82.0
python-dotenv==1.1.0
